refactor: Export structs, not interfaces
